Big changes!
Modified The Site so that it now sends a confirmation message using send_confirmation.py as well as a new listener which will run along with the game listener to see if any new numbers have responded with YES or STOP to opt in or out of messages.

This should make things better since otherwise anyone can just spam add numbers to be obnoxious.

Ideally I'll add a feature that limits the amount of times a phone number can be added and removed until it gets blacklisted.

** Note, all variables have been moved to the config/vars.py file, this allows me to reuse them them without copying a chunk of code everywhere!!

Signed-off-by: Eddie Federmeyer <eddiefedermeyer@yahoo.com>